HW2 - class name changed - toString method added - addAll method changed
